requirements.txt
# ── Backend (FastAPI + GCP + LangGraph) ──────────────────────
fastapi>=0.135.0
uvicorn[standard]>=0.43.0
python-dotenv>=1.0.0
langgraph>=1.1.0
langchain-core>=1.2.0
pydantic>=2.7.0
google-genai>=1.0.0
google-cloud-storage>=2.16.0
python-multipart>=0.0.9
httpx>=0.27.0

# ── ML Intelligence (Nodes 2 & 3 + NAT client) ──────────────
requests>=2.31.0
jsonschema>=4.21.0
pyyaml>=6.0.0

# ── Testing ──────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
